requirements.txt
flask
flask-cors
Pillow
diffusers
transformers
torch
torchvision
accelerate
